requirements.txt
Flask==1.1.1
tensorflow-cpu==2.4.0
mistune==0.8.4
numpy==1.19.5
Werkzeug==1.0.1
Pillow==8.1.0
tensorflow==2.4.1
gunicorn==19.3.0
wheel==0.35